Added a lot more cross references and lookup ability.
Can do quick lookup for master/copy parent/child shell/element relationships

The guid system is a lot better

Ironed out some bugs from earlier I did not catch

Added in the start of the messaging system, with the events able to tell when things changed by using sha1 references. Next commit or two this will be fully installed, and then all that is left is the gui for editing and display. The logic and the nitty gritty will be done behind the scenes as the gui can use what is here for full functionality of this demo