Version 0.3 (#2)
* removed some console.debug

* Tags can be Added and removed

* Adding tags is done by either modifying a tag array in the yaml , next to the game (or adding one if missing).
  Or by adding tags in the tag input at the top

* When adding editing the yaml , the tags will be synced when the yaml is loaded

* added in tags to example yaml

* Updated GUI to allow more controls

Better positioning of the tags (will replace that with a select2)

Added select 2 and first select2 for the elements (in the new column) with the data adapter to let me update the list easily

* Running shells, and the elements inside of them, now have guids

Needed to for controls to keep track of changes. Once the guid is set, it does not change for the running thing

Added UUID library to generate non repeating ids

* Added element_end to the shell element options

Can set the element end to be void, where nothing happens when a shell is popped, or return, where the same element name is sought out in the ancestor chain and all its variables are updated to that popped element's variables

Can now pop a shell, and the shell will be removed from the running shells, its elements might write values based on the element_end being return or not

* Added a lot more cross references and lookup ability.

Can do quick lookup for master/copy parent/child shell/element relationships

The guid system is a lot better

Ironed out some bugs from earlier I did not catch

Added in the start of the messaging system, with the events able to tell when things changed by using sha1 references. Next commit or two this will be fully installed, and then all that is left is the gui for editing and display. The logic and the nitty gritty will be done behind the scenes as the gui can use what is here for full functionality of this demo

* Finished the last two support functions for the gui

Next step is to hook up the emerging gui to this code. The rest of the work should be front end now, for this version

* WIP did the hooks onto existing gui, need to test everything next, then can add new stuff

* Fixed up new callback system and changes, after I retested

Broke up different gui things in new files, as now they do not have to be glomped in together

Debugged changes in the innermost layer (again)

Now I can add the new gui stuff

* updated readme with changes

* Stared on the edit element dialog box

Have a dropdown that displays a popup form when an element name is clicked

In the backend code, I added support for in place adding, editing and deleting of elements. So now, just need to build up gui things for display and edit of the element on display

* Got the view element dialog box mostly up, working on glom traces from the element

see the this-task todo

* The elements can now be added, edited and deleted via helpful GUI

The whole program is debugged, and is ready to add in the shell editing gui, which will be a lot of copy of the element, but a bit different

After that, just need to do a dot graph of the running, and make room for the display, and this is done

* WIP better looking gui on all screen sizes / major headway on shell editing

In the middle of code to edit shells in place, outside of the running environment

* Can now add/edit/remove master shells and the running shells will be updated accordingly

Fixed up and ran code for all of it

Editing is done, now can work on the presentation layer where the dot graph is shown , and will have ability to insert children to specific running shells, and to pop specific running shells

* Made room for the presentation part, the framework looks good on all screen sizes

* Figuring out dots...

* Now Renders dots with shells, elements and connections with gloms

Next Step is to put in click handlers for each shell and element to know which thing I am selecting

* Can select on dot and then push and pop

Added selection shell event, added helper functions, and dialog to push (place)

BUGFIX Fixed found element copy not getting its own guid, which was resulting in the push (place) not working well

Need to test for deeply nested shells, then can update the readme and release this version

* Glom lines now same color as their originating element

Fixed some bugs highlighted by the new version of phpstorm

* Added ability to edit running element variables from the gui

Fixed bug in the element shell edit list

Fixed bug in the shell edit parent list